ucastnici/tomas_bartko/lekce_8/requirements.txt
selenium==4.23.0
webdriver-manager==4.0.2
pytest==8.3.2
pytest-html==4.1.1
allure-pytest==2.13.2